requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi[standard]
uvicorn
groq
openai
setuptools
langchain-huggingface
langchain-core
huggingface_hub
transformers
chromadb
sentence-transformers
pymupdf
langchain
streamlit